% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascribe-package.R
\docType{package}
\name{ascribe-package}
\alias{ascribe}
\alias{ascribe-package}
\title{ascribe: Static AST Scanning of R Package and Function Usage}
\description{
Scans R source files for package and function use, then builds citations from configurable package universes. Supports .R, .Rmd, and .qmd files, resolves unqualified calls by attachment order and re-export origin, and leaves each package collection to define its own citations. See 'stanflow' \url{https://github.com/VisruthSK/stanflow} for an example usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ascribe.visruth.com}
  \item Report bugs at \url{https://github.com/VisruthSK/ascribe/issues}
}

}
\author{
\strong{Maintainer}: Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID}) [copyright holder]
}

}
\keyword{internal}
