% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{build_export_index}
\alias{build_export_index}
\title{Build an inverted export index}
\usage{
build_export_index(exports)
}
\arguments{
\item{exports}{Named list. Names are package names, values are
character vectors of function names.}
}
\value{
Named list mapping function names to character vectors of
package names.
}
\description{
Given a named list mapping package names to character vectors of
function names (as produced by \code{\link[=collect_pkg_funs]{collect_pkg_funs()}}), creates an
inverted index mapping function names to character vectors of
packages that export them.
}
\examples{
exports <- list(
  pkgA = c("foo", "bar"),
  pkgB = c("foo", "baz")
)
build_export_index(exports)
}
