% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{scan_skip_dirs}
\alias{scan_skip_dirs}
\title{Get directories to skip when scanning for R files}
\usage{
scan_skip_dirs()
}
\value{
Character vector of directory names.
}
\description{
Returns a character vector of directory names that should be
skipped during recursive directory scanning.
}
