% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{stdlib_funs}
\alias{stdlib_funs}
\title{Get standard library functions to ignore during scanning}
\usage{
stdlib_funs()
}
\value{
Character vector of function names.
}
\description{
Returns a character vector of base R function names that are
commonly used and typically not cited.
}
